LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/sandbox_024-expected.txt


PASS document.cookie access is allowed inside iframe with sandbox='allow-same-origin'.
